Enable unit testing of the DF API (#262)
