+++ .vscode/extensions.json
@@ -1,3 +1,8 @@
 {
-    "recommendations": ["editorconfig.editorconfig", "eg2.vscode-npm-script", "esbenp.prettier-vscode"]
+  "recommendations": [
+    "ms-azuretools.vscode-azurefunctions",
+    "editorconfig.editorconfig",
+    "eg2.vscode-npm-script",
+    "esbenp.prettier-vscode"
+  ]
 }
