+++ package.json
@@ -35,7 +35,6 @@
         "e2etst": "npm run "
     },
     "dependencies": {
-        "@azure/functions": "^1.0.2-beta2",
         "@types/lodash": "^4.14.119",
         "@types/uuid": "~3.4.4",
         "@types/validator": "^9.4.3",
@@ -49,6 +48,7 @@
         "validator": "~10.8.0"
     },
     "devDependencies": {
+        "@azure/functions": "^1.2.3",
         "@types/chai": "~4.1.6",
         "@types/chai-as-promised": "~7.1.0",
         "@types/chai-string": "~1.4.1",
@@ -73,7 +73,7 @@
         "prettier": "^2.0.5",
         "prettier-eslint": "11.0.0",
         "sinon": "~7.1.1",
-        "ts-node": "~1.0.0",
+        "ts-node": "^10.0.0",
         "typescript": "^3.8.3"
     },
     "engines": {
