+++ samples/UnitTesting/.funcignore
@@ -1,0 +1,7 @@
+*.js.map
+*.ts
+.git*
+.vscode
+local.settings.json
+test
+tsconfig.json
