+++ samples/UnitTesting/.mocharc.json
@@ -1,0 +1,6 @@
+{
+    "extension": ["ts"],
+    "spec": ["test/**/*.spec.ts"],
+    "package": "./package.json",
+    "r": ["esm", "ts-node/register", "source-map-support/register"]
+  }
