+++ samples/UnitTesting/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "unittesting",
+  "version": "1.0.0",
+  "description": "",
+  "scripts": {
+    "build": "tsc",
+    "watch": "tsc -w",
+    "prestart": "npm run build",
+    "start": "func start",
+    "test": "mocha"
+  },
+  "dependencies": {
+    "durable-functions": "^1.4.6",
+    "esm": "^3.2.25",
+    "mocha": "^8.4.0",
+    "sinon": "^10.0.0",
+    "source-map-support": "^0.5.19",
+    "ts-node": "^9.1.1"
+  },
+  "devDependencies": {
+    "@azure/functions": "^1.2.3",
+    "typescript": "^4.0.0"
+  }
+}
