+++ src/durableorchestrationcontext.ts
@@ -1,11 +1,73 @@
-import { EntityId, ITaskMethods, RetryOptions, Task, TimerTask } from "./classes";
+import { AggregatedError } from "./aggregatederror";
 import { TokenSource } from "./tokensource";
+import { DurableError } from "./durableerror";
+import {
+    EntityId,
+    ITaskMethods,
+    RetryOptions,
+    Task,
+    TimerTask,
+    CallActivityAction,
+    CallActivityWithRetryAction,
+    CallEntityAction,
+    CallHttpAction,
+    CallSubOrchestratorAction,
+    CallSubOrchestratorWithRetryAction,
+    ContinueAsNewAction,
+    CreateTimerAction,
+    DurableHttpRequest,
+    EventRaisedEvent,
+    EventSentEvent,
+    ExternalEventType,
+    GuidManager,
+    HistoryEvent,
+    HistoryEventType,
+    RequestMessage,
+    ResponseMessage,
+    SubOrchestrationInstanceCompletedEvent,
+    SubOrchestrationInstanceCreatedEvent,
+    SubOrchestrationInstanceFailedEvent,
+    TaskCompletedEvent,
+    TaskFactory,
+    TaskFailedEvent,
+    TaskFilter,
+    TaskScheduledEvent,
+    TimerCreatedEvent,
+    TimerFiredEvent,
+    WaitForExternalEventAction,
+} from "./classes";
+import { CompletedTask, TaskBase } from "./tasks/taskinterfaces";
 
 /**
  * Parameter data for orchestration bindings that can be used to schedule
  * function-based activities.
  */
 export class DurableOrchestrationContext {
+    constructor(
+        state: HistoryEvent[],
+        instanceId: string,
+        currentUtcDateTime: Date,
+        isReplaying: boolean,
+        parentInstanceId: string | undefined,
+        input: unknown
+    ) {
+        this.state = state;
+        this.instanceId = instanceId;
+        this.isReplaying = isReplaying;
+        this.currentUtcDateTime = currentUtcDateTime;
+        this.parentInstanceId = parentInstanceId;
+        this.input = input;
+
+        this.newGuidCounter = 0;
+        this.subOrchestratorCounter = 0;
+    }
+
+    private input: unknown;
+    private readonly state: HistoryEvent[];
+    private newGuidCounter: number;
+    private subOrchestratorCounter: number;
+    public customStatus: unknown;
+
     /**
      * The ID of the current orchestration instance.
      *
@@ -54,7 +116,71 @@ export class DurableOrchestrationContext {
      * Methods to handle collections of pending actions represented by [[Task]]
      * instances. For use in parallelization operations.
      */
-    public Task: ITaskMethods;
+    public Task: ITaskMethods = {
+        all: (tasks: TaskBase[]) => {
+            let maxCompletionIndex: number | undefined;
+            const errors: Error[] = [];
+            const results: Array<unknown> = [];
+            for (const task of tasks) {
+                if (!TaskFilter.isCompletedTask(task)) {
+                    return TaskFactory.UncompletedTaskSet(tasks);
+                }
+
+                if (!maxCompletionIndex) {
+                    maxCompletionIndex = task.completionIndex;
+                } else if (maxCompletionIndex < task.completionIndex) {
+                    maxCompletionIndex = task.completionIndex;
+                }
+
+                if (TaskFilter.isFailedTask(task)) {
+                    errors.push(task.exception);
+                } else {
+                    results.push(task.result);
+                }
+            }
+
+            // We are guaranteed that maxCompletionIndex is not undefined, or
+            // we would have alreayd returned an uncompleted task set.
+            const completionIndex = maxCompletionIndex as number;
+
+            if (errors.length > 0) {
+                return TaskFactory.FailedTaskSet(
+                    tasks,
+                    completionIndex,
+                    new AggregatedError(errors)
+                );
+            } else {
+                return TaskFactory.SuccessfulTaskSet(tasks, completionIndex, results);
+            }
+        },
+
+        any: (tasks: Task[]) => {
+            if (!tasks || tasks.length === 0) {
+                throw new Error("At least one yieldable task must be provided to wait for.");
+            }
+
+            let firstCompleted: CompletedTask | undefined;
+            for (const task of tasks) {
+                if (TaskFilter.isCompletedTask(task)) {
+                    if (!firstCompleted) {
+                        firstCompleted = task;
+                    } else if (task.completionIndex < firstCompleted.completionIndex) {
+                        firstCompleted = task;
+                    }
+                }
+            }
+
+            if (firstCompleted) {
+                return TaskFactory.SuccessfulTaskSet(
+                    tasks,
+                    firstCompleted.completionIndex,
+                    firstCompleted
+                );
+            } else {
+                return TaskFactory.UncompletedTaskSet(tasks);
+            }
+        },
+    };
 
     /**
      * Schedules an activity function named `name` for execution.
@@ -66,7 +192,35 @@ export class DurableOrchestrationContext {
      * function completes or fails.
      */
     public callActivity(name: string, input?: unknown): Task {
-        throw new Error("This is a placeholder.");
+        const newAction = new CallActivityAction(name, input);
+
+        const taskScheduled = this.findTaskScheduled(this.state, name);
+        const taskCompleted = this.findTaskCompleted(this.state, taskScheduled);
+        const taskFailed = this.findTaskFailed(this.state, taskScheduled);
+        this.setProcessed([taskScheduled, taskCompleted, taskFailed]);
+
+        if (taskCompleted) {
+            const result = this.parseHistoryEvent(taskCompleted);
+
+            return TaskFactory.SuccessfulTask(
+                newAction,
+                result,
+                taskCompleted.Timestamp,
+                taskCompleted.TaskScheduledId,
+                this.state.indexOf(taskCompleted)
+            );
+        } else if (taskFailed) {
+            return TaskFactory.FailedTask(
+                newAction,
+                taskFailed.Reason,
+                taskFailed.Timestamp,
+                taskFailed.TaskScheduledId,
+                this.state.indexOf(taskFailed),
+                new DurableError(taskFailed.Reason)
+            );
+        } else {
+            return TaskFactory.UncompletedTask(newAction);
+        }
     }
 
     /**
@@ -79,7 +233,96 @@ export class DurableOrchestrationContext {
      * function.
      */
     public callActivityWithRetry(name: string, retryOptions: RetryOptions, input?: unknown): Task {
-        throw new Error("This is a placeholder.");
+        const newAction = new CallActivityWithRetryAction(name, retryOptions, input);
+
+        let attempt = 1;
+        let taskScheduled: TaskScheduledEvent | undefined;
+        let taskFailed: TaskFailedEvent | undefined;
+        let taskRetryTimer: TimerCreatedEvent | undefined;
+        for (let i = 0; i < this.state.length; i++) {
+            const historyEvent = this.state[i];
+            if (historyEvent.IsProcessed) {
+                continue;
+            }
+
+            if (!taskScheduled) {
+                if (historyEvent.EventType === HistoryEventType.TaskScheduled) {
+                    if ((historyEvent as TaskScheduledEvent).Name === name) {
+                        taskScheduled = historyEvent as TaskScheduledEvent;
+                    }
+                }
+                continue;
+            }
+
+            if (historyEvent.EventType === HistoryEventType.TaskCompleted) {
+                if (
+                    (historyEvent as TaskCompletedEvent).TaskScheduledId === taskScheduled.EventId
+                ) {
+                    const taskCompleted = historyEvent as TaskCompletedEvent;
+                    this.setProcessed([taskScheduled, taskCompleted]);
+                    const result = this.parseHistoryEvent(taskCompleted);
+                    return TaskFactory.SuccessfulTask(
+                        newAction,
+                        result,
+                        taskCompleted.Timestamp,
+                        taskCompleted.TaskScheduledId,
+                        i
+                    );
+                } else {
+                    continue;
+                }
+            }
+
+            if (!taskFailed) {
+                if (historyEvent.EventType === HistoryEventType.TaskFailed) {
+                    if (
+                        (historyEvent as TaskFailedEvent).TaskScheduledId === taskScheduled.EventId
+                    ) {
+                        taskFailed = historyEvent as TaskFailedEvent;
+                    }
+                }
+                continue;
+            }
+
+            if (!taskRetryTimer) {
+                if (historyEvent.EventType === HistoryEventType.TimerCreated) {
+                    taskRetryTimer = historyEvent as TimerCreatedEvent;
+                } else {
+                    continue;
+                }
+            }
+
+            if (historyEvent.EventType === HistoryEventType.TimerFired) {
+                if ((historyEvent as TimerFiredEvent).TimerId === taskRetryTimer.EventId) {
+                    const taskRetryTimerFired = historyEvent as TimerFiredEvent;
+                    this.setProcessed([
+                        taskScheduled,
+                        taskFailed,
+                        taskRetryTimer,
+                        taskRetryTimerFired,
+                    ]);
+                    if (attempt >= retryOptions.maxNumberOfAttempts) {
+                        return TaskFactory.FailedTask(
+                            newAction,
+                            taskFailed.Reason,
+                            taskFailed.Timestamp,
+                            taskFailed.TaskScheduledId,
+                            i,
+                            new DurableError(taskFailed.Reason)
+                        );
+                    } else {
+                        attempt++;
+                        taskScheduled = undefined;
+                        taskFailed = undefined;
+                        taskRetryTimer = undefined;
+                    }
+                } else {
+                    continue;
+                }
+            }
+        }
+
+        return TaskFactory.UncompletedTask(newAction);
     }
 
     /**
@@ -90,8 +333,38 @@ export class DurableOrchestrationContext {
      * @param operationName The name of the operation.
      * @param operationInput The input for the operation.
      */
-    public callEntity(entityId: EntityId, operationName: string, operationInput?: unknown): Task {
-        throw new Error("This is a placeholder.");
+    public callEntity(entityId: EntityId, operationName: string, operationInput: unknown): Task {
+        const newAction = new CallEntityAction(entityId, operationName, operationInput);
+
+        const schedulerId = EntityId.getSchedulerIdFromEntityId(entityId);
+        const eventSent = this.findEventSent(this.state, schedulerId, "op");
+        let eventRaised;
+        if (eventSent) {
+            const eventSentInput =
+                eventSent && eventSent.Input
+                    ? (JSON.parse(eventSent.Input) as RequestMessage)
+                    : undefined;
+            eventRaised = eventSentInput
+                ? this.findEventRaised(this.state, eventSentInput.id)
+                : undefined;
+        }
+        this.setProcessed([eventSent, eventRaised]);
+
+        if (eventRaised) {
+            const parsedResult = this.parseHistoryEvent(eventRaised) as ResponseMessage;
+
+            return TaskFactory.SuccessfulTask(
+                newAction,
+                JSON.parse(parsedResult.result),
+                eventRaised.Timestamp,
+                eventSent.EventId,
+                this.state.indexOf(eventRaised)
+            );
+        }
+
+        // TODO: error handling
+
+        return TaskFactory.UncompletedTask(newAction);
     }
 
     /**
@@ -105,7 +378,55 @@ export class DurableOrchestrationContext {
      * the format `<calling orchestrator instance ID>:<#>`
      */
     public callSubOrchestrator(name: string, input?: unknown, instanceId?: string): Task {
-        throw new Error("This is a placeholder.");
+        if (!name) {
+            throw new Error(
+                "A sub-orchestration function name must be provided when attempting to create a suborchestration"
+            );
+        }
+
+        const newAction = new CallSubOrchestratorAction(name, instanceId, input);
+        const subOrchestratorCreated = this.findSubOrchestrationInstanceCreated(
+            this.state,
+            name,
+            instanceId
+        );
+        const subOrchestratorCompleted = this.findSubOrchestrationInstanceCompleted(
+            this.state,
+            subOrchestratorCreated
+        );
+        const subOrchestratorFailed = this.findSubOrchestrationInstanceFailed(
+            this.state,
+            subOrchestratorCreated
+        );
+
+        this.setProcessed([
+            subOrchestratorCreated,
+            subOrchestratorCompleted,
+            subOrchestratorFailed,
+        ]);
+
+        if (subOrchestratorCompleted) {
+            const result = this.parseHistoryEvent(subOrchestratorCompleted);
+
+            return TaskFactory.SuccessfulTask(
+                newAction,
+                result,
+                subOrchestratorCompleted.Timestamp,
+                subOrchestratorCompleted.TaskScheduledId,
+                this.state.indexOf(subOrchestratorCompleted)
+            );
+        } else if (subOrchestratorFailed) {
+            return TaskFactory.FailedTask(
+                newAction,
+                subOrchestratorFailed.Reason,
+                subOrchestratorFailed.Timestamp,
+                subOrchestratorFailed.TaskScheduledId,
+                this.state.indexOf(subOrchestratorFailed),
+                new DurableError(subOrchestratorFailed.Reason)
+            );
+        } else {
+            return TaskFactory.UncompletedTask(newAction);
+        }
     }
 
     /**
@@ -124,7 +445,112 @@ export class DurableOrchestrationContext {
         input?: unknown,
         instanceId?: string
     ): Task {
-        throw new Error("This is a placeholder.");
+        if (!name) {
+            throw new Error(
+                "A sub-orchestration function name must be provided when attempting to create a suborchestration"
+            );
+        }
+
+        const newAction = new CallSubOrchestratorWithRetryAction(
+            name,
+            retryOptions,
+            input,
+            instanceId
+        );
+
+        let attempt = 1;
+        let subOrchestratorCreated: SubOrchestrationInstanceCreatedEvent | undefined;
+        let subOrchestratorFailed: SubOrchestrationInstanceFailedEvent | undefined;
+        let taskRetryTimer: TimerCreatedEvent | undefined;
+        for (let i = 0; i < this.state.length; i++) {
+            const historyEvent = this.state[i];
+            if (historyEvent.IsProcessed) {
+                continue;
+            }
+
+            if (!subOrchestratorCreated) {
+                if (historyEvent.EventType === HistoryEventType.SubOrchestrationInstanceCreated) {
+                    const subOrchEvent = historyEvent as SubOrchestrationInstanceCreatedEvent;
+                    if (
+                        subOrchEvent.Name === name &&
+                        (!instanceId || instanceId === subOrchEvent.InstanceId)
+                    ) {
+                        subOrchestratorCreated = subOrchEvent;
+                    }
+                }
+                continue;
+            }
+
+            if (historyEvent.EventType === HistoryEventType.SubOrchestrationInstanceCompleted) {
+                if (
+                    (historyEvent as SubOrchestrationInstanceCompletedEvent).TaskScheduledId ===
+                    subOrchestratorCreated.EventId
+                ) {
+                    const subOrchCompleted = historyEvent as SubOrchestrationInstanceCompletedEvent;
+                    this.setProcessed([subOrchestratorCreated, subOrchCompleted]);
+                    const result = this.parseHistoryEvent(subOrchCompleted);
+                    return TaskFactory.SuccessfulTask(
+                        newAction,
+                        result,
+                        subOrchCompleted.Timestamp,
+                        subOrchCompleted.TaskScheduledId,
+                        i
+                    );
+                } else {
+                    continue;
+                }
+            }
+
+            if (!subOrchestratorFailed) {
+                if (historyEvent.EventType === HistoryEventType.SubOrchestrationInstanceFailed) {
+                    if (
+                        (historyEvent as SubOrchestrationInstanceFailedEvent).TaskScheduledId ===
+                        subOrchestratorCreated.EventId
+                    ) {
+                        subOrchestratorFailed = historyEvent as SubOrchestrationInstanceFailedEvent;
+                    }
+                }
+                continue;
+            }
+
+            if (!taskRetryTimer) {
+                if (historyEvent.EventType === HistoryEventType.TimerCreated) {
+                    taskRetryTimer = historyEvent as TimerCreatedEvent;
+                }
+                continue;
+            }
+
+            if (historyEvent.EventType === HistoryEventType.TimerFired) {
+                if ((historyEvent as TimerFiredEvent).TimerId === taskRetryTimer.EventId) {
+                    const taskRetryTimerFired = historyEvent as TimerFiredEvent;
+                    this.setProcessed([
+                        subOrchestratorCreated,
+                        subOrchestratorFailed,
+                        taskRetryTimer,
+                        taskRetryTimerFired,
+                    ]);
+                    if (attempt >= retryOptions.maxNumberOfAttempts) {
+                        return TaskFactory.FailedTask(
+                            newAction,
+                            subOrchestratorFailed.Reason,
+                            subOrchestratorFailed.Timestamp,
+                            subOrchestratorFailed.TaskScheduledId,
+                            i,
+                            new DurableError(subOrchestratorFailed.Reason)
+                        );
+                    } else {
+                        attempt += 1;
+                        subOrchestratorCreated = undefined;
+                        subOrchestratorFailed = undefined;
+                        taskRetryTimer = undefined;
+                    }
+                } else {
+                    continue;
+                }
+            }
+        }
+
+        return TaskFactory.UncompletedTask(newAction);
     }
 
     /**
@@ -139,7 +565,41 @@ export class DurableOrchestrationContext {
         headers?: { [key: string]: string },
         tokenSource?: TokenSource
     ): Task {
-        throw new Error("This is a placeholder");
+        if (content && typeof content !== "string") {
+            content = JSON.stringify(content);
+        }
+
+        const req = new DurableHttpRequest(method, uri, content as string, headers, tokenSource);
+        const newAction = new CallHttpAction(req);
+
+        // callHttp is internally implemented as a well-known activity function
+        const httpScheduled = this.findTaskScheduled(this.state, "BuiltIn::HttpActivity");
+        const httpCompleted = this.findTaskCompleted(this.state, httpScheduled);
+        const httpFailed = this.findTaskFailed(this.state, httpScheduled);
+        this.setProcessed([httpScheduled, httpCompleted, httpFailed]);
+
+        if (httpCompleted) {
+            const result = this.parseHistoryEvent(httpCompleted);
+
+            return TaskFactory.SuccessfulTask(
+                newAction,
+                result,
+                httpCompleted.Timestamp,
+                httpCompleted.TaskScheduledId,
+                this.state.indexOf(httpCompleted)
+            );
+        } else if (httpFailed) {
+            return TaskFactory.FailedTask(
+                newAction,
+                httpFailed.Reason,
+                httpFailed.Timestamp,
+                httpFailed.TaskScheduledId,
+                this.state.indexOf(httpFailed),
+                new DurableError(httpFailed.Reason)
+            );
+        } else {
+            return TaskFactory.UncompletedTask(newAction);
+        }
     }
 
     /**
@@ -148,7 +608,9 @@ export class DurableOrchestrationContext {
      * @param The JSON-serializable data to re-initialize the instance with.
      */
     public continueAsNew(input: unknown): Task {
-        throw new Error("This is a placeholder.");
+        const newAction = new ContinueAsNewAction(input);
+
+        return TaskFactory.UncompletedTask(newAction);
     }
 
     /**
@@ -166,7 +628,22 @@ export class DurableOrchestrationContext {
      * @returns A TimerTask that completes when the durable timer expires.
      */
     public createTimer(fireAt: Date): TimerTask {
-        throw new Error("This is a placeholder.");
+        const newAction = new CreateTimerAction(fireAt);
+
+        const timerCreated = this.findTimerCreated(this.state, fireAt);
+        const timerFired = this.findTimerFired(this.state, timerCreated);
+        this.setProcessed([timerCreated, timerFired]);
+
+        if (timerFired) {
+            return TaskFactory.CompletedTimerTask(
+                newAction,
+                timerFired.Timestamp,
+                timerFired.TimerId,
+                this.state.indexOf(timerFired)
+            );
+        } else {
+            return TaskFactory.UncompletedTimerTask(newAction);
+        }
     }
 
     /**
@@ -174,7 +651,7 @@ export class DurableOrchestrationContext {
      * value.
      */
     public getInput<T>(): T {
-        throw new Error("This is a placeholder.");
+        return this.input as T;
     }
 
     /**
@@ -186,8 +663,12 @@ export class DurableOrchestrationContext {
      * this value is a combination of the orchestration instance ID and an
      * internally managed sequence number.
      */
-    public newGuid(): string {
-        throw new Error("This is a placeholder.");
+    public newGuid(instanceId: string): string {
+        const guidNameValue = `${instanceId}_${this.currentUtcDateTime.valueOf()}_${
+            this.newGuidCounter
+        }`;
+        this.newGuidCounter++;
+        return GuidManager.createDeterministicGuid(GuidManager.UrlNamespaceValue, guidNameValue);
     }
 
     /**
@@ -204,7 +685,7 @@ export class DurableOrchestrationContext {
      * orchestrator function's custom status.
      */
     public setCustomStatus(customStatusObject: unknown): void {
-        throw new Error("This is a placeholder.");
+        this.customStatus = customStatusObject;
     }
 
     /**
@@ -215,6 +696,254 @@ export class DurableOrchestrationContext {
      * using [[raiseEvent]].
      */
     public waitForExternalEvent(name: string): Task {
-        throw new Error("This is a placeholder.");
+        const newAction = new WaitForExternalEventAction(name, ExternalEventType.ExternalEvent);
+
+        const eventRaised = this.findEventRaised(this.state, name);
+        this.setProcessed([eventRaised]);
+
+        if (eventRaised) {
+            const result = this.parseHistoryEvent(eventRaised);
+
+            return TaskFactory.SuccessfulTask(
+                newAction,
+                result,
+                eventRaised.Timestamp,
+                eventRaised.EventId,
+                this.state.indexOf(eventRaised)
+            );
+        } else {
+            return TaskFactory.UncompletedTask(newAction);
+        }
+    }
+
+    // ===============
+    /* Returns undefined if not found. */
+    private findEventRaised(state: HistoryEvent[], eventName: string): EventRaisedEvent {
+        const returnValue = eventName
+            ? state.filter((val: HistoryEvent) => {
+                  return (
+                      val.EventType === HistoryEventType.EventRaised &&
+                      (val as EventRaisedEvent).Name === eventName &&
+                      !val.IsProcessed
+                  );
+              })[0]
+            : undefined;
+        return returnValue as EventRaisedEvent;
+    }
+
+    /* Returns undefined if not found. */
+    private findEventSent(
+        state: HistoryEvent[],
+        instanceId: string,
+        eventName: string
+    ): EventSentEvent {
+        const returnValue = eventName
+            ? state.filter((val: HistoryEvent) => {
+                  return (
+                      val.EventType === HistoryEventType.EventSent &&
+                      (val as EventSentEvent).InstanceId === instanceId &&
+                      (val as EventSentEvent).Name === eventName &&
+                      !val.IsProcessed
+                  );
+              })[0]
+            : undefined;
+        return returnValue as EventSentEvent;
+    }
+
+    /* Returns undefined if not found. */
+    private findSubOrchestrationInstanceCreated(
+        state: HistoryEvent[],
+        name: string,
+        instanceId: string | undefined
+    ): SubOrchestrationInstanceCreatedEvent | undefined {
+        const matches = state.filter((val: HistoryEvent) => {
+            return (
+                val.EventType === HistoryEventType.SubOrchestrationInstanceCreated &&
+                !val.IsProcessed
+            );
+        });
+
+        if (matches.length === 0) {
+            return undefined;
+        }
+
+        this.subOrchestratorCounter++;
+
+        // Grab the first unprocessed sub orchestration creation event and verify that
+        // it matches the same function name and instance id if provided. If not, we know that
+        // we have nondeterministic behavior, because the callSubOrchestrator*() methods were not
+        // called in the same order this replay that they were scheduled in.
+        const returnValue = matches[0] as SubOrchestrationInstanceCreatedEvent;
+        if (returnValue.Name !== name) {
+            throw new Error(
+                `The sub-orchestration call (n = ${this.subOrchestratorCounter}) should be executed with a function name of ${returnValue.Name} instead of the provided function name of ${name}. Check your code for non-deterministic behavior.`
+            );
+        }
+
+        if (instanceId && returnValue.InstanceId !== instanceId) {
+            throw new Error(
+                `The sub-orchestration call (n = ${this.subOrchestratorCounter}) should be executed with an instance id of ${returnValue.InstanceId} instead of the provided instance id of ${instanceId}. Check your code for non-deterministic behavior.`
+            );
+        }
+        return returnValue;
+    }
+
+    /* Returns undefined if not found. */
+    private findSubOrchestrationInstanceCompleted(
+        state: HistoryEvent[],
+        createdSubOrch: SubOrchestrationInstanceCreatedEvent | undefined
+    ): SubOrchestrationInstanceCompletedEvent | undefined {
+        if (createdSubOrch === undefined) {
+            return undefined;
+        }
+
+        const matches = state.filter((val: HistoryEvent) => {
+            return (
+                val.EventType === HistoryEventType.SubOrchestrationInstanceCompleted &&
+                (val as SubOrchestrationInstanceCompletedEvent).TaskScheduledId ===
+                    createdSubOrch.EventId &&
+                !val.IsProcessed
+            );
+        });
+
+        return matches.length > 0
+            ? (matches[0] as SubOrchestrationInstanceCompletedEvent)
+            : undefined;
+    }
+
+    /* Returns undefined if not found. */
+    private findSubOrchestrationInstanceFailed(
+        state: HistoryEvent[],
+        createdSubOrchInstance: SubOrchestrationInstanceCreatedEvent | undefined
+    ): SubOrchestrationInstanceFailedEvent | undefined {
+        if (createdSubOrchInstance === undefined) {
+            return undefined;
+        }
+
+        const matches = state.filter((val: HistoryEvent) => {
+            return (
+                val.EventType === HistoryEventType.SubOrchestrationInstanceFailed &&
+                (val as SubOrchestrationInstanceFailedEvent).TaskScheduledId ===
+                    createdSubOrchInstance.EventId &&
+                !val.IsProcessed
+            );
+        });
+
+        return matches.length > 0 ? (matches[0] as SubOrchestrationInstanceFailedEvent) : undefined;
+    }
+
+    /* Returns undefined if not found. */
+    private findTaskScheduled(state: HistoryEvent[], name: string): TaskScheduledEvent | undefined {
+        const returnValue = name
+            ? (state.filter((val: HistoryEvent) => {
+                  return (
+                      val.EventType === HistoryEventType.TaskScheduled &&
+                      (val as TaskScheduledEvent).Name === name &&
+                      !val.IsProcessed
+                  );
+              })[0] as TaskScheduledEvent)
+            : undefined;
+        return returnValue;
+    }
+
+    /* Returns undefined if not found. */
+    private findTaskCompleted(
+        state: HistoryEvent[],
+        scheduledTask: TaskScheduledEvent | undefined
+    ): TaskCompletedEvent | undefined {
+        if (scheduledTask === undefined) {
+            return undefined;
+        }
+
+        const returnValue = scheduledTask
+            ? (state.filter((val: HistoryEvent) => {
+                  return (
+                      val.EventType === HistoryEventType.TaskCompleted &&
+                      (val as TaskCompletedEvent).TaskScheduledId === scheduledTask.EventId
+                  );
+              })[0] as TaskCompletedEvent)
+            : undefined;
+        return returnValue;
+    }
+
+    /* Returns undefined if not found. */
+    private findTaskFailed(
+        state: HistoryEvent[],
+        scheduledTask: TaskScheduledEvent | undefined
+    ): TaskFailedEvent | undefined {
+        if (scheduledTask === undefined) {
+            return undefined;
+        }
+
+        const returnValue = scheduledTask
+            ? (state.filter((val: HistoryEvent) => {
+                  return (
+                      val.EventType === HistoryEventType.TaskFailed &&
+                      (val as TaskFailedEvent).TaskScheduledId === scheduledTask.EventId
+                  );
+              })[0] as TaskFailedEvent)
+            : undefined;
+        return returnValue;
+    }
+
+    /* Returns undefined if not found. */
+    private findTimerCreated(state: HistoryEvent[], fireAt: Date): TimerCreatedEvent {
+        const returnValue = fireAt
+            ? state.filter((val: HistoryEvent) => {
+                  return (
+                      val.EventType === HistoryEventType.TimerCreated &&
+                      new Date((val as TimerCreatedEvent).FireAt).getTime() === fireAt.getTime()
+                  );
+              })[0]
+            : undefined;
+        return returnValue as TimerCreatedEvent;
+    }
+
+    /* Returns undefined if not found. */
+    private findTimerFired(
+        state: HistoryEvent[],
+        createdTimer: TimerCreatedEvent | undefined
+    ): TimerFiredEvent | undefined {
+        const returnValue = createdTimer
+            ? (state.filter((val: HistoryEvent) => {
+                  return (
+                      val.EventType === HistoryEventType.TimerFired &&
+                      (val as TimerFiredEvent).TimerId === createdTimer.EventId
+                  );
+              })[0] as TimerFiredEvent)
+            : undefined;
+        return returnValue;
+    }
+
+    private setProcessed(events: Array<HistoryEvent | undefined>): void {
+        events.map((val: HistoryEvent | undefined) => {
+            if (val) {
+                val.IsProcessed = true;
+            }
+        });
+    }
+
+    private parseHistoryEvent(directiveResult: HistoryEvent): unknown {
+        let parsedDirectiveResult: unknown;
+
+        switch (directiveResult.EventType) {
+            case HistoryEventType.EventRaised:
+                const eventRaised = directiveResult as EventRaisedEvent;
+                parsedDirectiveResult =
+                    eventRaised && eventRaised.Input ? JSON.parse(eventRaised.Input) : undefined;
+                break;
+            case HistoryEventType.SubOrchestrationInstanceCompleted:
+                parsedDirectiveResult = JSON.parse(
+                    (directiveResult as SubOrchestrationInstanceCompletedEvent).Result
+                );
+                break;
+            case HistoryEventType.TaskCompleted:
+                parsedDirectiveResult = JSON.parse((directiveResult as TaskCompletedEvent).Result);
+                break;
+            default:
+                break;
+        }
+
+        return parsedDirectiveResult;
     }
 }
